requirements.txt
fastapi
uvicorn
PyPDF2
python-dotenv
pydantic
python-multipart
openai>=1.0.0
